[realm] extract packet buffering to trait + cleanup network code
